Fix library install_name on Mac OS X
On Mac OS X, each shared library has an “install_name” property that
should be set to the absolute path of the library's installation
location, since this is how OS X find libraries at launch time.  In
waf, this is easy enough — just include the “osx” tool into your
wscript.

Unfortunately, this breaks waf unit testing [1], since you can longer
run the test executables directly from the build directory.  Normally,
the test framework sets DYLD_LIBRARY_PATH so that the test executables
find the newly compiled libraries in the build directory.  However,
the install path includes the full three-digit vnum of the shared
library, while the file produced in the build directory doesn't.  (At
install time, the no-vnum file is renamed into the with-vnum file.)

In the waf bug [1], we've submitted a patch that fixes this, which is
included in a new waf script in this project template.  It creates an
additional symlink in the build directory that contains the
three-digit vnum, so that the test executables can find their
libraries.

[1] http://code.google.com/p/waf/issues/detail?id=721