Always look for libcheck
Before, we only looked for the libcheck library in our test wscript
file, since the main libraries and programs don't link with it.
Unfortunately, the test directory isn't included unless the user is
running "waf build_tests" or "waf test".  It doesn't get included
for "waf configure".  That causes link errors when building our tests.
By looking for libcheck in the src directory, too -- even though it's
never used there -- we ensure that it's available for the test
directory.