C99 fixed-size integer types
The libcork/core/types.h file will ensure that several standard C99
types are available.  For now, we just assume that we're using an actual
C99-compliant compiler and libc, so we just include the appropriate
standard header files.  If we need to support other systems in the
future, we'll add the appropriate preprocessor fu to define typedefs for
the C99 types.  We also have a test case that ensures that all of the
fixed-size types are available.