Preprocessor sizeof macros
This patch adds several macros that make the size of various types
available to the preprocessor.  Our current implementation assumes that
we have the C99 integer type macros available.  To find the size of int,
for instance, we just compare INT_MAX against INT8_MAX and friends.
There's also a macro for the size of a pointer; our implementation
assumes that intptr_t and uintptr_t are the same size as void *, so we
perform the same test using INTPTR_MAX.  The test case verifies that the
sizes that we calculate in the preprocessor match the sizes returned by
the sizeof operator.