Endianness macros
This patch adds a slew of macros for dealing with endianness issues.  We
can currently detect the host's endianness on Linux and Mac OS X.  (We
need to know which OS to be able to know which flavor of endian.h is
available.)  We provide our own detection macros — CORK_HOST_ENDIANNESS
and CORK_OTHER_ENDIANNESS — as well as C string equivalents.  There is
also a collection of macros for swapping unsigned integer values, both
in place via a statement-like macro, or as the result of a function-like
macro.  There are variants for converting between host-endianness and
either big- or little-endianness, as well as variants for performing an
explicit byte swap regardless of the underlying endianness.