Test both shared and embedded libcork
In the README file, we mention that there are two ways to use libcork:
either link with a shared libcork, which is installed by our build
scripts, or embed the whole thing into your own project.  This patch
updates our testing framework to test both of these uses.  That means
that we create two versions of each test case.  For the embedded
version, we now have to mention which submodules the test case needs, so
that we don't needlessly recompile the whole of libcork for each test
case.

Right now, there's only a single submodule, called “core”, and all of
our code is header-only, so there's no real difference between the
shared and embedded versions.  But that should start to change as we add
more features.