Added doxygen config file
This patch adds a doxygen configuration file.  It also updates the
existing source files to organize the generated documentation correctly.
For now, you run doxygen by hand, from the root of the source tree:

    $ doxygen docs/Doxyfile

That will generate HTML documentation in docs/html.  The git repo is set
to ignore this output directory.