Managed binary buffers and slices
This patch adds an API for “managed” binary buffers.  A managed buffer
wraps a binary buffer with a reference count and a finalization closure.
Code access the buffer through slices, which point to a subset of the
managed buffer.  As slices are created and destroyed, they update the
reference count of the managed buffer that they point into.  Once there
are no remaining slices that point into a managed buffer, the buffer is
destroyed.