Resizable buffer class
This patch adds a new resizable buffer class.  This is different than
the “managed” buffer class; this class isn't reference counted, since
it's intended to only have a single owner at a time.  This buffer class
is mutable; there are methods for changing the buffer's contents.
(Though we need to flesh these out with additional methods.)  There are
also methods for changing a mutable, resizable buffer into an immutable,
managed buffer.  Note that once you create a managed buffer, the
contents become immutable, so that there's less to worry about when
passing the data between threads.