Stream consumer interface
The new cork_stream_consumer_t interface lets you process a stream of
binary data.  The data is delivered to the consumer via the interface's
“data” method, one chunk at a time.  Each chunk is represented by a
cork_slice_t instance.  Once the producer has provided all of the data,
it calls the interface's “eof” method, which gives the consumer a chance
to raise an error condition at the end of the stream, if necessary.

This patch adds an implementation of the interface that appends all of
the data into a cork_buffer_t instance.