Configuration module
Several libcork features (will) have different implementations on
different platforms.  One of the design decisions we've made is to use
the C preprocessor to autodetect which platform we're on, and which
implementations to select.  That makes the library easier to embed into
projects with a wide variety of build systems.

This patch adds a new Configuration module, which is where we'll place
all of that auto-detection logic.  That way it's not sprinkled
throughout all of our other, more interesting, header files.  The
Configuration module defines a set of configuration macros, which our
autodetection logic will ensure have been defined.  Any other part of
libcork can just include <libcork/config.h> to have access to these
macros, and use them (instead of their own platform detection) to choose
which implementation to select.

Currently, the only thing that needs to be auto-detected is the
endianness of the system; we have different detection logic on Linux and
Mac OS X.