Static library to emulate embedding libcork
One of the use cases we want to support is embedding libcork within
another application or library, without any of libcork's symbols being
exported publicly.  We compile each of our test cases twice — once
linking with the shared libcork, and once “embedding” libcork into the
test case's executable.  Before, we recompiled the libcork sources
separately for each embedded test case.  This patch updates things so
that we compile a static version of libcork, and link that in to each
embedded test case.  That still simulates directly compiling libcork
into the executables, without all of the recompilation.