Hash tables
This patch adds a new hash table data structure.  The implementation is
based on the public domain st.[hc] files written by Peter Moore at UCB.
There are two methods for adding values to the hash table: one which
takes in a pre-initialized value, and overwrites it if there's already
an entry with the given key; and one which only takes in the key, and
either returns the existing value, or creates a new, empty value, if the
key isn't already present.