Error reporting
This patch adds functions and macros for reporting error conditions.
These are kind of like exceptions, but not as heavy-weight.  For
instance, there's no automatic support for propagating these error
conditions up the call chain; you have to do this yourself.

The API is modeled after glib's GError API.  The main difference is that
with GError, the error objects are allocated on the heap, whereas we
allocate them on the stack.  There is also not yet any support for
printf-like format strings when setting an error; for now, you can only
provide a single string.

Each error condition is represented by a “class” and a “code”.  Classes
serve as namespaces for errors; this gets around the problem that
errno-based solutions have, where there's no easy mechanism for new
modules to add new error codes without clashing.  Codes are simple
integers, and only have to be unique within a class; you're free to
assign them as you wish.  (The common solution is to use an enum.)
Classes must be globally unique; unlike GError, which uses quarks
generated at runtime as class IDs, we use precomputed hash values.  This
is similar to using a random value, but the value is produced from
applying a hash function to some string that identifies the namespace of
error conditions.

To illustrate all of this, the IP address class now reports parse errors
when you try to parse an address from a string.  The error class
(defined by CORK_NET_ADDRESS_ERROR) is a hash of the name of the
corresponding include file.