printf-life buffer methods
This patch adds several additional methods to the cork_buffer_t class,
which let you set or append to the buffer contents using a printf-like
format string.  There are printf and vprintf versions for each, which
let you provide the values for the format string either as variadic
function parameters, or as a stdarg.h va_list instance.

In addition, the cork_error_set function has been updated to take in a
format string and values, rather than just a bare string.