Hierarchical allocator
This patch adds a hierarchical allocator, modeled after the halloc [1]
and talloc [2] projects.  Each object is associated with a parent;
freeing any object also causes all of its children to be freed.  We
support destructor functions, which are called just before the memory
for an object is freed.  We don't support cycles or additional
references to objects, yet.

[1] http://swapped.cc/halloc/
[2] http://ccan.ozlabs.org/info/talloc.html