Slices no longer tied to managed buffers
This patch updates the cork_slice_t class to work with other kinds of
underlying buffer, instead of being specifically tied to a
cork_managed_buffer_t.  I've also split the slice code off into
separate header and source files.

In addition, slices now support two important operations.  The old
“slice” method has been renamed to “copy”, since it's used to initialize
a new slice to point at the same underlying buffer.  Note that this
doesn't necessarily mean that the underlying buffer is copied, just that
there's a copy of the original slice.  If your buffer is smart (like
cork_managed_buffer_t), then this can involve a reference count
increment and no actual copying of the data.

The second operation is called “slice”.  It updates an existing slice to
point at a subset of its contents, without affecting the underlying
buffer.  Nothing is freed or copied; in most cases, this simply requires
updating the buf and size fields.