Clean up managed buffer class
I've reorganized the managed buffer class a bit to make it cleaner for
subclasses.  In particular, we don't store the libcork allocator in the
managed buffer anymore; if subclasses need that, they can store it
themselves.  And the free function has been moved into an interface
struct.  For a one-method interface like this, it doesn't save any space
(and in fact, adds an extra indirection), but it seems good to be
consistent with the other classes.