Timestamp type
This patch adds a new timestamp type, whose representation follows that
used by the NTP protocol.  Timestamps are represented by 64-bit
quantities, and each unit of timestamp represents 2^-32 seconds.  With
this representation, the upper 32 bits represents the seconds portion of
the timestamp, and timestamps have a resolution of about 233
picoseconds and a range of about 136 years.