Changing naming convention for types
This patch changes our naming convention for libcork types.  Before, we
would use typedefs to create a bunch of “cork_foo_t” types.  Now, we
leave out the typedefs, and require you to use the “struct” and “enum”
tags explicitly.  (The “_t” suffix is technically reserved by the POSIX
standard for system types.)