Reference counting garbage collector
This patch adds an implementation of the reference counting garbage
collector described in §3 of [1].  The garbage collector correctly works
in the presence of reference cycles.

We only implement the sequential garbage collector; we assume that each
object is only used within a single thread at a time.  Eventually we'll
implement a cork_gc_migrate function that will let you safely migrate an
object from one thread to another.

[1] Bacon, DF and Rajan, VT. Concurrent cycle collection in reference
    counted systems. Proc. ECOOP 2001. LNCS 2072.