cork_strdup and cork_strfree functions
This patch adds a cork_alloc equivalent to the standard strdup function.
The cork_strfree function must be used to free a string created using
cork_strdup.  We have to store the size of the string along with the
copied contents, since we'll need that when we free the string in
cork_strfree.