GC free function receives GC object, not allocator
Each garbage-collected object can define a free function, which is
called when the object has been collected by the garbage collector.  It
allows you to perform any additional cleanup before the object is fully
destroyed.  Before, we would pass in the garbage collector's allocator
instance; now, we pass in the garbage collector object itself.  The GC
object contains a link to the allocator, so no capabilities have been
lost.  And often, the GC object will be embedded in some larger state
object, and this gives the free function access to that state object.