Adding Sphinx documentation for config and core
This patch moves all of the Doxygen documentation into a new Sphinx
document collection.  This makes it a lot easier to write the
documentation as running text, that happens to include a reference entry
for each type and function, rather than as an API reference that happens
to have explanatory text thrown in in places.