Better error API
This patch updates the cork_error API.  Before, the human-readable
description of an error was stored directly in the cork_error struct.
Now, we instead store a pointer to a function that can generate the
description later on.  This eliminates some expensive printf calls if
they're not going to be used.  It also should make it easier to localize
error messages, since that just implies a more complex printer function.
Error descriptions can still include arbitrary data, but this must be
copied into the cork_error struct so that it's available when the
printer function is called.