Better handling of NUL terminators in cork_buffer
We now always append a NUL terminator onto the end of the contents of a
cork_buffer, but don't include the NUL byte in the length of the buffer.
This simplifies a couple of the functions, and is “saner”, in that its
description in the documentation is much shorter.  One backwards-
incompatible change is that the "size" of a C string buffer is now one
less, since it no longer contains the NUL terminator.