Resizable arrays
This patch implements a new cork_array family of types.  cork_array is
parameterized on the array's element type.  The array class's mutator
functions automatically resize the array's storage as needed to hold all
of the array's elements.  In addition, we reserve a small amount of
storage in the array itself, which lets us avoid one heap allocation
when storing a small number of elements.