Build error strings in cork_error_set again
Well that was a fun experiment.  It turns out that it's much simpler to
provide the error message using printf-like parameters in
cork_error_set, rather than deferring the message formatting into a
separate function.  (Not the least because all of those string
references will often go away before the printer function is called...)
These error messages are still localizable; you'll just need to use a
gettext macro right when you call cork_error_set.

If, later on, we still want to allow callers to check the error class
and code, without going through the effort of formatting the
description, then we do that with some kind of flag inside the
cork_error struct.