CORK_LIKELY and CORK_UNLIKELY
This patch adds macros that let you tell the compiler whether a Boolean
expression is likely to be true or false.  The compiler can sometimes
use this information to generate more efficient code.  The
error-checking macros now use CORK_UNLIKELY to indicate that the error
conditions are not the likely outcome.