cork_slice_init_static
The new cork_slice_init_static function lets you initialize a slice
object to point at a static buffer.  You can also use it to wrap a
non-static buffer, as long as you can ensure that buffer exists for at
least as long as the slice and any copies made of the slice.