CORK_ATTR_NOINLINE macro
The CORK_ATTR_NOINLINE macro lets you tell the compiler that a function
shouldn't be considered for inlining.  We use that attribute in the
memory pool implementation to make the fast path of cork_mempool_new as
lean as possible.  (Inlining cork_mempool_new_block makes us spill over
into the stack more, and so there's additional prologue and epilogue
code.  We only want to incur the stack cost if we actually have to
allocate a block.)