Inline cork_mempool_new
The cork_mempool_new function is now defined in the public header file
as an inlinable static function.  The slow path, where we allocate and
initialize a new block, is still defined as an actual function in the
library.  But the fast path can now be incorporated into the caller for
speed.