Memory allocation that aborts on failure
This patch adds cork_{m,c,re}alloc macros that allocate memory and abort
the program if they fail.  This seems to be a more standard error
recovery mechanism, since it's not like we're going to be able to do any
real recovery anyway.  There are also cork_x{m,c,re}alloc macros that
still return NULL on allocation failure, in case you really want to try
to recover.  All memory allocation that libcork does itself uses the
abort-on-failure versions.  The error-checking helper macros still
implement recover-on-failure, but they're not as useful anymore.