Don't use slices in stream consumer API
You now use a raw pointer and size_t to pass in the next chunk of data
to a cork_stream_consumer.  This is a simpler API, and most stream
consumers that need to copy the data will end up doing something more
complex than a simple cork_slice_copy.