Remove cork_tls infinite loop in Release builds
The code in the pthread implementation of the cork_tls() macro was
incorrectly putting some function calls inside of an assert check.  In
Release builds, the assert calls are left out of the compiled code, so
these function calls weren't being performed.  This led to compilation
errors (due to unused symbols) and an infinite loop when initializing
the thread-local storage.