cork_once allows recursion
The cork_once macro now works properly when the initialization function
being called is recursive.  When a thread detects that the function is
already being executed, it checks to see if it's the thread doing that
execution.  If so, it immediately returns, to allow the recursion to
continue.  Otherwise, it waits for the other thread to finish the
initialization.