Test option parsing in CLI programs
We didn't have a test case that checked a subcommand that uses getopt.
It turns out the getopt expects an argv that's structured like you get
in your main() function — you have to have the “executable name” appear
as its first element.  The command driver code ensures this is the case
now; for nested command sets, the executable name will be the name of
the command set.