Portable thread IDs
This patch updates the `cork_thread_id` type to have a portable
definition that's consistent across all platforms.  That means that we
must assign the thread IDs ourselves, since we can't rely on
`pthread_get_self` (or its equivalent) to get the thread ID.

We assign IDs by maintaining a thread-local descriptor structure.
`cork_tls` initializes this structure with 0 bytes when it's first
allocated or accessed, which corresponds to the "no thread ID assigned"
value for our new `cork_thread_id` type.  This lets us detect that we
haven't assigned an ID to the current thread yet, increment our thread
ID counter, and return that newly incremented value.