Macro versions of several IP address functions
Several of the IP address functions have been turned into macros, to
help inline their simple definitions when they're used.  The function
versions are still available, but with a trailing underscore in their
name.  Note that this change will introduce an ABI (but not API)
incompatibility in the next release.