Power-of-2 number of hash table bins
We now ensure that the number of hash table bins is a power of 2.
Before, we were ensuring that it was a prime number.  This was useful
for old-school hash functions, which only provided uniform distribution
modulo a prime number of bins.  We're using a nicer hash function that
is uniform regardless, and switching to a power-of-2 number of bins lets
us use a tasty fast logical AND instead of a modulo operator.