“Stable” and “big” hash values
This patch introduces three new functions for calculating hash values.
A “stable” hash is guaranteed to be the same regardless of which
platform the code is running on.  It's also guaranteed to be consistent
across libcork versions.  A “big” hash value is drawn from a much larger
space of possible hashes (128 bits vs 32), giving you a correspondingly
smaller chance of hash collisions.  We still use MurmurHash3 as the
implementation of all three hash variants.