Propagate termination signals to child processes
The `cork_subprocess_start_and_wait` function now catches the `SIGHUP`,
`SIGINT`, and `SIGTERM` signals, and propagate them on to child
processes.  If any of the signals is currently being ignored by the
process, they will still be ignored.  (The glibc documentation claims
that this is important to interact nicely with “non-job-control
shells”.)

The documentation used to say that this could be controlled on a
per-subprocess basis using a flag, but the way I've actually implemented
it is to always propagate those signals on.