Capture stdout and stderr of subprocesses
We now actually do something with the stream consumers that you pass in
to the `cork_subprocess_new_exec` function.  For any non-NULL consumers
that you provide, we create an unnamed pipe to capture the corresponding
output stream from the subprocess.  Instead of using a simple
`waitpid(3)` call to wait for all of the child subprocesses to exit, we
now use a `select(3)` loop to give us a chance to drain the stdout and
stderr pipes, passing any data that we read into the corresponding
stream consumer.