Controlled draining of subprocesses
You can now have more control over how and when we try to read data from
the stdin and stderr streams of a group of subprocesses.  We've added a
new “subprocess group” object that contains the state that carries over
between function calls.  We still need a global reference to the current
subprocess group for our signal handlers, and so we're still not
thread-safe.