Symbol visibility macros
This patch adds the new `CORK_EXPORT`, `CORK_IMPORT`, and `CORK_LOCAL`
macros, which can be used to declare the visibility of each function in
a library.  There is also documentation describing how to use these
macros; you usually won't use `CORK_EXPORT` and `CORK_IMPORT` directly,
but instead use them to define a `FOO_API` macro for your library.

Our “embedded” copy of libcork, which we statically link with each of
our test case programs, defines `CORK_API` to be `CORK_LOCAL`, which
seems to let the compiler apply a couple of extra optimizations.  (The
linked programs are all smaller.)