Process cleanup functions
You can now register functions that will be automatically called when
the current process exits, similar to the standard `atexit(3)` function.
Our version differs in that you must provide a priority value for each
function, allowing you to explicitly define the order in which the
cleanup functions should be called.