Fix alignment in cork_array on 32-bit platforms
On 32-bit platforms, our internal representation of a cork_array (which
doesn't depend on the types of the elements) wasn't aligned correctly
for all element types.  We can grab the pointer to the internal array in
a slightly different way, however, which works for all element types
regardless of alignment.