File and directories
This patch adds a new `cork_file` type to represent a file or directory
in the local filesystem.  Its API is modeled on combination of the Java
`File` class and the Python `os` and `os.path` modules.

We currently have several functions that operate on files.  One tests
whether a file exists, and if so, whether its a file, directory, or
symlink.  One removes a file or directory, and optionally its contents.
One creates a new directory, and optionally its parents.