Copy-once slices
A copy-once slice can be used to wrap an existing buffer, when you can
guarantee that the buffer will outlive the slice.  This means that we
don't need to make an initial copy of the buffer; we can just use the
existing buffer directly.  The first time we copy the slice, though, we
duplicate the original wrapped buffer into a new managed buffer
instance.  (We have to do this because we can't make any guarantees
about how long the copied slice will live.)  We also go ahead and update
the initial slice (which originally wrapped the existing buffer
directly) to point into the new managed buffer.  This ensures that if we
copy the original slice twice, we won't create yet another duplicate of
the underlying buffer.