Lightly copied slices
You can now make a “light” copy of a slice.  This is similar to the
existing “full” copy operation, in that the new slice points into the
same buffer as the original.  However, by making a light copy, you
assert that you'll ensure that the newly copied slice will not outlive
the original slice.  Some slice implementations can use that guarantee
to make a copy with less overhead.  For instance, the new copy-once
slice already knows that the underlying buffer will outlive the original
slice, and will therefore also outlive any light copies.  That lets us
reuse the existing buffer for light copies.  (For full copies, we can't
assume any guarantees about how the life of the underlying buffer
relates to the lives of the new full copies, and so we still have to
make duplicate the actual slice content.)