Allow custom library installation directory
It turns out that some Linux distros (ahem, RHEL) want you to put 64-bit
libraries into /usr/lib64, not /usr/lib.  So we need to allow the user
to override the destination directory that we use when installing
libraries.  This new `CMAKE_INSTALL_LIBDIR` variable seems to be the
standard one to use.  We cannot rely on it being either a relative or
absolute path (see [1]), but the usage in this patch should work either
way.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=795542