Other hash types in cork-hash
The cork-hash utility now takes in a handful of command-line options
that let you choose between the different kinds of hash value that we
can produce: stable (the default, reproducing the previous behavior),
fastest, and big.  Just like the underlying C functions, the final two
are not guaranteed to provide consistent results on different platforms,
or across different versions of libcork.