Delete known entry from hash table
The new `cork_hash_table_delete_entry` function lets you delete an entry
from hash table more efficiently.  You pass in a `cork_hash_table_entry`
instance, instead of a key value.  This means that we don't have to
search through the hash table for the value to delete; the entry
instance acts as a pointer to the specific entry that we want to delete.