Don't install empty man page directories
The CMake logic for installing man-page link files would create a
directory for each section that might possibly contain links.  Now it
only creates directories for the sections that actually contain links.
This also fixes a bug when using CMake 2.6, where the docs/links
directory needed to be present, even if your project has no links
whatsoever.  It's now fine for the directory to be missing.