Better rounding of timestamps
Rounding the fractional portion of a timestamp make involving adding a
carry bit to the seconds portion.  For instance, rounding 0.999999 to
milliseconds should result in 1.000.  That means that we can't separate
the act of extracting seconds and gammaseconds and then round to some
other fractional denominator.  So, the functions for rounding to ms, µs,
or ns now return both the seconds portions and the desired fractional
portion.

In addition, there are now two accessors for extracting the seconds
portion on its own.  The existing `cork_timestamp_sec` truncates towards
zero.  The (probably more useful) `cork_timestamp_round_sec` rounds to
the nearest second.

This also means that the `%f` format specifier is no longer valid, since
the width that you select for the fractional portion also affects the
value that we render for the seconds portion.  So instead of a separate
`%f` fractional specifier, the `%s` and `%S` specifiers now accept
optional width specifiers to include the fractional seconds in the
output.